fqdn/messagehandler: remove FQDN suite
There's no sense in creating a suite just for a single benchmark. Remove
it.
